requirements.txt
# Core: FAISS, embeddings, DataFrames
faiss-cpu>=1.10.0
numpy>=2.0
pandas>=2.0
torch>=2.0
PyYAML>=6.0
python-dotenv>=0.19.0
google-genai>=1.45.0
transformers>=4.40.0
tqdm>=4.65.0

# Optional: pull job postings from CockroachDB / Postgres-compatible (read_data.py)
beautifulsoup4>=4.13.0
sqlalchemy>=2.0.0
sqlalchemy-cockroachdb>=2.0.0
psycopg2-binary>=2.9.0

# Notebooks / eval
scikit-learn>=1.3.0

# Scripts under model tests/
openai>=1.0.0
huggingface_hub>=0.20.0
